factor out some migration helper methods and shut ganache logging up during testing
