+++ packages/core/lib/commands/migrate/index.js
@@ -1,6 +1,8 @@
 module.exports = {
   run: require("./run"),
   meta: require("./meta"),
+  runMigrations: require("./runMigrations"),
+  setUpDryRunEnvironmentThenRunMigrations: require("./setUpDryRunEnvironmentThenRunMigrations"),
   determineDryRunSettings: require("./determineDryRunSettings"),
   prepareConfigForRealMigrations: require("./prepareConfigForRealMigrations")
 };
