+++ packages/core/lib/commands/migrate/run.js
@@ -1,90 +1,43 @@
 module.exports = async function (options) {
-  const Artifactor = require("@truffle/artifactor");
-  const Resolver = require("@truffle/resolver");
-  const Migrate = require("@truffle/migrate");
   const WorkflowCompile = require("@truffle/workflow-compile");
   const { Environment } = require("@truffle/environment");
   const Config = require("@truffle/config");
-  const { promisify } = require("util");
-  const copy = require("../../copy");
   const determineDryRunSettings = require("./determineDryRunSettings");
   const prepareConfigForRealMigrations = require("./prepareConfigForRealMigrations");
+  const runMigrations = require("./runMigrations");
+  const setUpDryRunEnvironmentThenRunMigrations = require("./setUpDryRunEnvironmentThenRunMigrations");
   const tmp = require("tmp");
   tmp.setGracefulCleanup();
 
-  const conf = Config.detect(options);
-  if (conf.compileNone || conf["compile-none"]) {
-    conf.compiler = "none";
+  const config = Config.detect(options);
+  if (config.compileNone || config["compile-none"]) {
+    config.compiler = "none";
   }
 
-  const result = await WorkflowCompile.compileAndSave(conf);
-  await WorkflowCompile.assignNames(conf, result);
-  await Environment.detect(conf);
+  const result = await WorkflowCompile.compileAndSave(config);
+  await WorkflowCompile.assignNames(config, result);
+  await Environment.detect(config);
 
   const { dryRunOnly, dryRunAndMigrations } = determineDryRunSettings(
-    conf,
+    config,
     options
   );
 
   if (dryRunOnly) {
-    conf.dryRun = true;
-    await setupDryRunEnvironmentThenRunMigrations(conf);
+    config.dryRun = true;
+    await setUpDryRunEnvironmentThenRunMigrations(config);
   } else if (dryRunAndMigrations) {
-    const currentBuild = conf.contracts_build_directory;
-    conf.dryRun = true;
+    const currentBuild = config.contracts_build_directory;
+    config.dryRun = true;
 
-    await setupDryRunEnvironmentThenRunMigrations(conf);
+    await setUpDryRunEnvironmentThenRunMigrations(config);
 
-    let { config, proceed } = await prepareConfigForRealMigrations(
+    const { preparedConfig, proceed } = await prepareConfigForRealMigrations(
       currentBuild,
       options
     );
-    if (proceed) await runMigrations(config);
+    if (proceed) await runMigrations(preparedConfig);
   } else {
-    await runMigrations(conf);
-  }
-
-  async function setupDryRunEnvironmentThenRunMigrations(config) {
-    await Environment.fork(config, {
-      logging: {
-        quiet: true
-      },
-      // we need to tell Ganache to not unlock any accounts so that only
-      // user's accounts are unlocked since this will be a dry run
-      wallet: {
-        totalAccounts: 0
-      }
-    });
-    // Copy artifacts to a temporary directory
-    const temporaryDirectory = tmp.dirSync({
-      unsafeCleanup: true,
-      prefix: "migrate-dry-run-"
-    }).name;
-
-    await promisify(copy)(config.contracts_build_directory, temporaryDirectory);
-
-    config.contracts_build_directory = temporaryDirectory;
-    // Note: Create a new artifactor and resolver with the updated config.
-    // This is because the contracts_build_directory changed.
-    // Ideally we could architect them to be reactive of the config changes.
-    config.artifactor = new Artifactor(temporaryDirectory);
-    config.resolver = new Resolver(config);
-
-    return await runMigrations(config);
-  }
-
-  async function runMigrations(config) {
-    if (options.f) {
-      return await Migrate.runFrom(options.f, config);
-    } else {
-      const needsMigrating = await Migrate.needsMigrating(config);
-
-      if (needsMigrating) {
-        return await Migrate.run(config);
-      } else {
-        config.logger.log("Network up to date.");
-        return;
-      }
-    }
+    await runMigrations(config);
   }
 };
